style: Norminette and split into multiple files
Also add lexer.h and add some definitions used in lexer to defines.h.
