+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/08 15:43:22 by lyeh              #+#    #+#             */
-/*   Updated: 2023/12/17 20:47:50 by ldulling         ###   ########.fr       */
+/*   Updated: 2023/12/20 10:52:12 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,6 @@
 
 bool	ft_init_shell(t_shell *shell, char **env);
 bool	mock_lexer(t_shell *shell);
-
-/* Lexer */
-bool	ft_lexer(t_shell *shell);
 
 char	*ft_get_token_type_str(int type);
